+++ drivers/net/ethernet/marvell/octeontx2/nic/otx2_txrx.h
@@ -76,7 +76,7 @@ struct otx2_rcv_queue {
 
 struct sg_list {
 	u16	num_segs;
-	u64	skb;
+	uintptr_t	skb;
 	u64	size[OTX2_MAX_FRAGS_IN_SQE];
 	u64	dma_addr[OTX2_MAX_FRAGS_IN_SQE];
 };
